feat: [PQ-210] Logic updates.
